# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Rest interactiva y pública para el proyecto Anirank.",
        "title": "Anirank API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte API Anirank",
            "email": "admin@anirank.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/playlists": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new playlist for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlists"
                ],
                "summary": "Create Playlist",
                "parameters": [
                    {
                        "description": "Playlist Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "description": {
                                    "type": "string"
                                },
                                "is_public": {
                                    "type": "boolean"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/domain.Playlist"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.AppError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.AppError"
                        }
                    }
                }
            }
        },
        "/playlists/users/{id}": {
            "get": {
                "description": "Fetch paginated playlists for a user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Playlists"
                ],
                "summary": "List User Playlists",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Playlist"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "err": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Artist": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "name_jp": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Badge": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.ExternalLink": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domain.Format": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Genre": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Playlist": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_public": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.PlaylistSong"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.PlaylistSong": {
            "type": "object",
            "properties": {
                "playlist_id": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "song": {
                    "description": "Enriched",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Song"
                        }
                    ]
                },
                "song_id": {
                    "type": "integer"
                }
            }
        },
        "domain.Post": {
            "type": "object",
            "properties": {
                "anilist_id": {
                    "type": "integer"
                },
                "banner": {
                    "type": "string"
                },
                "banner_src": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_links": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ExternalLink"
                    }
                },
                "format": {
                    "$ref": "#/definitions/domain.Format"
                },
                "format_id": {
                    "type": "integer"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Genre"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "producers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Producer"
                    }
                },
                "season": {
                    "$ref": "#/definitions/domain.Season"
                },
                "season_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "songs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Song"
                    }
                },
                "status": {
                    "type": "boolean"
                },
                "studios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Studio"
                    }
                },
                "thumbnail": {
                    "type": "string"
                },
                "thumbnail_src": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "year": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Year"
                        }
                    ]
                },
                "year_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Role": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Season": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Song": {
            "type": "object",
            "properties": {
                "anime_id": {
                    "type": "integer"
                },
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Artist"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Post"
                        }
                    ]
                },
                "season": {
                    "$ref": "#/definitions/domain.Season"
                },
                "season_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "song_en": {
                    "type": "string"
                },
                "song_jp": {
                    "type": "string"
                },
                "song_romaji": {
                    "type": "string"
                },
                "theme_num": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.SongVariant"
                    }
                },
                "views": {
                    "type": "integer"
                },
                "year": {
                    "$ref": "#/definitions/domain.Year"
                },
                "year_id": {
                    "type": "integer"
                }
            }
        },
        "domain.SongVariant": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "embed_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "season": {
                    "$ref": "#/definitions/domain.Season"
                },
                "season_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "song": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Song"
                        }
                    ]
                },
                "song_id": {
                    "type": "integer"
                },
                "spoiler": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "version_number": {
                    "type": "integer"
                },
                "video_src": {
                    "description": "Embedded Video Info to simulate 1:1 legacy relation without join overhead",
                    "type": "string"
                },
                "video_type": {
                    "description": "file, embed",
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "year": {
                    "$ref": "#/definitions/domain.Year"
                },
                "year_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "badges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Badge"
                    }
                },
                "banner": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "roles": {
                    "description": "RBAC Relations",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Role"
                    }
                },
                "score_format": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.Year": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}